requirements.txt
flask
sqlalchemy
psycopg2-binary
pandas
python-dotenv
